# docs/openapi_full.yaml
openapi: 3.0.3
info:
  title: SOFATELCOM SOFATELCOM-V2 API (Full Spec)
  version: '1.0.0'
  description: |-
    Full API specification for SOFATELCOM app (web + mobile). Includes endpoints
    for interventions, dispatching, stock, reservations, RH, tracking, auth
    and exports. Mobile support includes token-based auth and tracking endpoints
servers:
  - url: http://localhost:4300
    description: Local development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
...
    Intervention:
      type: object
      properties:
        id:
          type: integer
        demande_id:
          type: integer
        technicien_id:
          type: integer
        statut:
          type: string
        date_debut:
          type: string
          format: date-time
        date_fin:
          type: string
          format: date-time
        gps_lat:
          type: string
        gps_long:
          type: string
        photos_list:
          type: array
          items:
            type: string
        signature_client:
          type: string
    InterventionCreate:
      type: object
      required: [demande_id, technicien_id]
      properties:
        demande_id:
          type: integer
        technicien_id:
          type: integer
        date_debut:
          type: string
          format: date-time
        gps_lat:
          type: string
        gps_long:
          type: string
        note:
          type: string
...
paths:
...
  /api/interventions:
    get:
      summary: List interventions (filterable)
      parameters:
        - name: date_debut
          in: query
          schema:
            type: string
          required: false
        - name: date_fin
          in: query
          schema:
            type: string
        - name: technicien_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: interventions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Intervention'
  /api/intervention:
    post:
      summary: Create an intervention (mobile / web)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterventionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Intervention'
        '400':
          $ref: '#/components/schemas/Error'
  /api/intervention/{intervention_id}:
    get:
      summary: Get details for an intervention
      parameters:
        - name: intervention_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Intervention'
    put:
      summary: Update an intervention
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterventionCreate'
      responses:
        '200':
          description: Updated
...
  /api/export/interventions:
    get:
      summary: Export interventions (CSV/Excel)
      parameters:
        - name: date_debut
          in: query
          schema:
            type: string
        - name: date_fin
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, excel]
      responses:
        '200':
          description: file download
...
security:
  - cookieAuth: []
  - bearerAuth: []